Add function to add reward to a user on compeleting a challenge.

+++ src/main/java/com/group02/mindmingle/controller/UserController.java
@@ -125,7 +125,7 @@ public class UserController {
     }
 
     @PostMapping("/user/rewards")
-    public ResponseEntity<String> addReward(@RequestParam long rewardId) {
+    public ResponseEntity<String> addReward(@RequestParam int rewardId) {
         userService.addReward(rewardId);
         return ResponseEntity.ok("Reward added successfully.");
     }
